Deactivate Angular, begin setting up Vue
