+++ package.json
@@ -53,6 +53,7 @@
     "bluebird": "^3.3.4",
     "font-awesome": "^4.5.0",
     "socket.io-client": "^1.4.5",
-    "vue": "^1.0.24"
+    "vue": "^1.0.24",
+    "vue-router": "^0.7.13"
   }
 }
