+++ src/ui/_app.ts
@@ -2,6 +2,8 @@
 ///<reference path="../vendor/requirejs.d.ts" />
 ///<reference path="../vendor/angular.d.ts" />
 ///<reference path="../vendor/jquery.d.ts" />
+///<reference path="../vendor/vue.d.ts" />
+///<reference path="../vendor/vue-router.d.ts" />
 ///<reference path="../vendor/bluebird.d.ts" />
 ///<reference path="../vendor/fetch.d.ts" />
 ///<reference path="../vendor/socket.io-client.d.ts" />
