+++ src/ui/index.html
@@ -11,7 +11,9 @@
 </head>
 
 <body>
-    <div class="content" ng-view></div>
+    <div class="content" id="content">
+        <router-view></router-view>
+    </div>
     <div class="draggable"></div>
 
     <div class="app-controls">
