+++ tsconfig.json
@@ -1,7 +1,8 @@
 {
 	"compilerOptions": {
 		"target": "es5",
-		"sourceMap": true
+		"sourceMap": true,
+		"experimentalDecorators": true
 	},
 	"exclude": [
 		"node_modules"
